requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
clu==0.0.3
contextlib2==0.6.0.post1
cycler==0.10.0
dill==0.3.3
flatbuffers==1.12
flax==0.3.3
future==0.18.2
gast==0.4.0
gin-config==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
importlib-metadata==4.0.1
importlib-resources==5.1.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
ml-collections==0.1.0
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.2.52
opt-einsum==3.3.0
Pillow==8.2.0
promise==2.3
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.3
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc0
tensorflow-datasets==4.3.0
tensorflow-hub==0.12.0
tensorflow-metadata==0.30.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021032501
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
bert-tensorflow==1.0.1

